+++ androidApp/src/main/java/hr/dtakac/prognoza/ui/forecast/ForecastDataList.kt
@@ -9,10 +9,15 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import hr.dtakac.prognoza.presentation.TextResource
 import hr.dtakac.prognoza.presentation.asString
-import hr.dtakac.prognoza.presentation.forecast.ForecastUi
+import hr.dtakac.prognoza.presentation.forecast.*
+import hr.dtakac.prognoza.shared.entity.Description
+import hr.dtakac.prognoza.shared.entity.Mood
 import hr.dtakac.prognoza.ui.common.keyVisibilityPercent
+import hr.dtakac.prognoza.ui.theme.AppTheme
 import hr.dtakac.prognoza.ui.theme.PrognozaTheme
 import hr.dtakac.prognoza.ui.theme.asWeatherIconResId
 import kotlinx.coroutines.flow.distinctUntilChanged
@@ -183,3 +188,104 @@ private fun ItemVisibilityChanges(
             }
     }
 }
+
+@Composable
+private fun TodayScreenPreview(
+    current: CurrentUi = fakeCurrentUi(),
+    today: TodayUi? = fakeTodayUi(),
+    coming: List<ComingDayUi>? = fakeComingUi()
+) = AppTheme(mood = Mood.CLEAR_DAY) {
+    ForecastDataList(
+        data = ForecastUi(
+            current = current,
+            today = today,
+            coming = coming
+        )
+    )
+}
+
+@Preview
+@Composable
+private fun AllVisiblePreview() = TodayScreenPreview()
+
+@Preview
+@Composable
+private fun TodayScreenJustCurrentPreview() = TodayScreenPreview(
+    today = null,
+    coming = null
+)
+
+@Preview
+@Composable
+private fun TodayScreenCurrentPlusTodayPreview() = TodayScreenPreview(
+    coming = listOf()
+)
+
+@Preview
+@Composable
+private fun TodayScreenCurrentPlusComingPreview() = TodayScreenPreview(
+    today = null
+)
+
+private fun fakeCurrentUi(): CurrentUi = CurrentUi(
+    place = TextResource.fromString("Tenja"),
+    date = TextResource.fromString("September 12"),
+    temperature = TextResource.fromString("1°"),
+    description = TextResource.fromString("Clear sky, sleet soon"),
+    weatherIconDescription = Description.CLEAR_SKY_DAY,
+    wind = TextResource.fromString("Wind: 15 km/h"),
+    feelsLike = TextResource.fromString("Feels like: -21°"),
+    mood = Mood.CLEAR_DAY,
+    precipitation = TextResource.fromString("12 mm")
+)
+
+private fun fakeTodayUi(): TodayUi = TodayUi(
+    lowHighTemperature = TextResource.fromString("135°—197°"),
+    hourly = mutableListOf<DayHourUi>().apply {
+        for (i in 1..4) {
+            add(
+                DayHourUi(
+                    time = TextResource.fromString("14:00"),
+                    temperature = TextResource.fromString("23°"),
+                    precipitation = TextResource.fromString("1.99 mm"),
+                    description = TextResource.fromString("Clear and some more text"),
+                    weatherIconDescription = if (i % 2 == 0) Description.CLEAR_SKY_DAY else Description.UNKNOWN
+                )
+            )
+        }
+    }
+)
+
+private fun fakeComingUi(): List<ComingDayUi> = listOf(
+    ComingDayUi(
+        date = TextResource.fromString("Thu, Sep 13"),
+        lowHighTemperature = TextResource.fromString("16—8"),
+        precipitation = TextResource.empty(),
+        hours = mutableListOf<ComingDayHourUi>().apply {
+            for (i in 1..12) {
+                add(
+                    ComingDayHourUi(
+                        time = TextResource.fromString("$i:00"),
+                        temperature = TextResource.fromString("20"),
+                        weatherIconDescription = Description.CLEAR_SKY_DAY
+                    )
+                )
+            }
+        },
+        weatherIconDescriptions = listOf()
+    ),
+    ComingDayUi(
+        date = TextResource.fromString("Fri, Sep 14"),
+        lowHighTemperature = TextResource.fromString("18—8"),
+        precipitation = TextResource.fromString("0.7 mm"),
+        hours = listOf(),
+        weatherIconDescriptions = listOf()
+    ),
+    ComingDayUi(
+        date = TextResource.fromString("Sat, Sep 15"),
+        lowHighTemperature = TextResource.fromString("21—5"),
+        precipitation = TextResource.empty(),
+        hours = listOf(),
+        weatherIconDescriptions = listOf()
+    )
+)